Add files via upload
GenerateDB creates the Project2 database and stored procedures.  GenerateTables creates the tables and procedures in the Project2 database.  PopulateTables seeds the database.  Output php files are the server-side processing of the MSSQL database, which displays on the procedure php screens via the ajax javascript file.  Styles.css is styling for the screens.